fix: password regex uniformity issue
