+++ web-app/src/app/types.ts
@@ -65,22 +65,6 @@ export type AppErrors = {
   [Property in ErrorSource]: AppError | null;
 };
 
-export const passwordValidationError = (
-  <div>
-    Password must
-    <ul style={{ marginTop: 0, paddingLeft: '15px' }}>
-      <li>Contain at least one uppercase letter</li>
-      <li>Contain at least one lowercase letter</li>
-      <li>Contain at least one digit</li>
-      <li>
-        Contain at least one special char
-        {'(^ $ * . [ ] { } ( ) ? " ! @ # % & / \\ , > < \' : ; | _ ~ `)'}
-      </li>
-      <li>Be at least 12 chars long</li>
-    </ul>
-  </div>
-);
-
 export enum OauthProvider {
   Google = 'Google',
   Github = 'Github',
